New warning level and renaming of some types

+++ src/Camera.cpp
@@ -22,10 +22,10 @@ int Camera::RenderCopy(SDL_Renderer* renderer, SDL_Texture* texture,
         return SDL_RenderCopy(renderer, texture, srcrect, NULL);
     }
     else {
-        transrect.x = (dstrect->x - x) * zoom + hscreenw;
-        transrect.y = (-dstrect->y + y) * zoom + hscreenh;
-        transrect.w = dstrect->w * zoom;
-        transrect.h = dstrect->h * zoom;
+        transrect.x = (int)((dstrect->x - x) * zoom) + hscreenw;
+        transrect.y = (int)((-dstrect->y + y) * zoom) + hscreenh;
+        transrect.w = (int)(dstrect->w * zoom);
+        transrect.h = (int)(dstrect->h * zoom);
 
         return SDL_RenderCopy(renderer, texture, srcrect, &transrect);
     }
@@ -41,10 +41,10 @@ int Camera::RenderFillRect(SDL_Renderer* renderer,
         return SDL_RenderFillRect(renderer, NULL);
     }
     else {
-        transrect.x = (rect->x - x) * zoom + hscreenw;
-        transrect.y = (-rect->y + y) * zoom + hscreenh;
-        transrect.w = rect->w * zoom;
-        transrect.h = rect->h * zoom;
+        transrect.x = (int)((rect->x - x) * zoom) + hscreenw;
+        transrect.y = (int)((-rect->y + y) * zoom) + hscreenh;
+        transrect.w = (int)(rect->w * zoom);
+        transrect.h = (int)(rect->h * zoom);
 
         return SDL_RenderFillRect(renderer, &transrect);
     }
@@ -57,7 +57,10 @@ int Camera::RenderDrawLine(SDL_Renderer* renderer,
                                double x2,
                                double y2) {
 
-    return SDL_RenderDrawLine(renderer, (x1-x)*zoom+hscreenw, (-y1+y)*zoom+hscreenh,
-                              (x2-x)*zoom+hscreenw, (-y2+y)*zoom+hscreenh);
+    return SDL_RenderDrawLine(renderer,
+                              (int)((x1-x) * zoom) + hscreenw,
+                              (int)((-y1+y) * zoom) + hscreenh,
+                              (int)((x2-x) * zoom) + hscreenw,
+                              (int)((-y2+y) * zoom) + hscreenh);
 
 }
